Added/modified bash formatted text printing functions and queue interaction aliases and minor ImageJ macro modifications
The primary substantive modifications were to the prompt appearance (PS1/PS2/PS3/PS4 variables). The pCT user profile defines a default prompt appearance, but the implementation of this has been moved up to the pct_user_script so users have easy access to this and can modify it as desired to produce the format/appearance they prefer. The previously hardcode text color/formatting of the PS1 prompt has also been replaced by insertion of various variable values, those related to both text color/format and those corresponding to the insertion of system/PWD properties (e.g., date/time, user/host name, git status (when applicable, etc.). Several unicode symbols were defined and atypically, 1 was added to the main prompt and among the available arrow symbols, 3 were selected to define the PS2/PS3/PS4 prompts while the others remain currently unused. With this modified implementation, users will find it quite easy to understand the current definition of these prompts and use this understanding to modify them as they desire.

Added and modified formatted/colored text printing functions loaded with the pCT user .bash_profile and sourced scripts. In particular, functions with text format specification arguments were expanded to support text color, background color, and text format specification integers as a single comma separated list rather than text color alone or separate arguments for each. These were also expanded to more easily allow "key=value" printing statements with independently controlled text formatting and the ability to specify the column where the "value" string is to appear, inserting spaces with either the key or value formatting as specified by the user by choosing the -a # or -A # function argument.

Other additions/modifications of the bash scripts were related to global file path/name definitions associated with the development of a standard reconstruction program testing routineand the definition of aliases for determining all reconstruction jobs a user currently has residing in the queue and for deleting all such jobs. A testing script was also added for bash code development, merely as a simplification of the process.

The ImageJ macro modifications were minor, most notably was a change such that the Test_Parameters_#.txt files are to be located in a single constant location rather than in the parent directory of the reconstruction output data to be analyzed. The other changes were modified printing statements and the addition/modification of the directory contents querying/printing functions and the basic/generic single line implementation functions (primarily those for printing). The functions for querying directory contents are now being expanded to support recursive queryies into available subdirectories as deeply as desired. The goal is to use this to be able to analyze all reconstructions within the target directory as well as any others that appear in some subdirectory of it.