Committing restructuring ImageJ storage and development preparing ROI_analysis for varying ROI selection types/shapes
Development towards extending ROI_analysis beyond exclusively circular ROI selections was performed in development_testbench.ijm. The variables/arrays used in the 2 primary macros (ROI_analysis/merge_data) and their control macros (<...>_control) remain without a completely consistent naming correspondence and a significant portion of this development included adding all namings of a partcicular assignment to both ROI_analysis and development_testbench so that they each contain every variable naming used in either macro. Constant value variables/arrays are intended to have upper case naming and many of the naming differences are lower/upper case naming conflicts, so a number of these were resolved in the process. Lower case, constant value variable/array assignments were removed from the macros when each usage of the lower case name in the macro was replaced, so the lower case names remaining in these macros correspond to variables/arrays which likely still contain assignments referring to this name somewhere in the macro, but this has been considerably reduced at this point.

The ROI definiition files for each phantom are now parsed using a slightly different scheme since the # of parameters required to make an ROI selection varies depending on the ROI selection type/shape (e.g. oval, rectangle, polygon,... have different # of points). Hence, parameters for each ROI are separated by a ';' and those for a particular ROI are now separated by a ','. The slices in which the ROI should be selected are now also included in this file instead of the analysis.txt/.cfg file since it is unlikely that each ROI is to be selected in the same set of slices as the others so the set of slices an ROI selection is to be made needs to be specified for each ROI explicitly using the same ';'/',' separation scheme.

Several helper functions were added to aid in the development of this variable ROI shape selection and analysis scheme and a few existing functions were modified so their capabilities could be extended to support this new scheme. A new feature added in this development is the option to fill in each ROI selection and write the corresponding slice to PNG so an image is obtained that shows the ROIs in which analyses were performed. The ROI_diameter/ROI_selection_radii/ROI_profile_radius_beyond_ROI fields were retained without modification to retain support for the usage of the original "circle" ROI selection and a reformulation of the original routine for such ROIs is performed when this ROI selection is present. However, an equivalent selection process can be performed with an "oval" selection, so support for the original scheme will likely be dropped. This will eliminate the need for the ROI_selection_radii field and the ROI_diameter/ROI_profile_radius_beyond_ROI fields will likely be reformulated as a separate ROI selection type option (e.g. "line_profile" w/ ROI_diameter as a parameter in 1 of the 2 CSV lists).

This new ROI selection analysis scheme, as implemented in development_testbench, has now been tested and validated. The new/modified variables, arrays, and functions have now been added to ROI_analysis and for safety, this commit  includes all necessary additions/modifications prior to replacement of the original ROI analysis scheme with the new scheme (i.e. ROI_analysis is still configured to perform the original analysis). The nesting of the ROI selection analysis within slice #/diameter #/ROI #/etc loops was roughly/approximately duplicated within the development macro so this nesting and slice/diameter/ROI iteration scheme could be retained entirely without modification and the process for adding this new ROI selection analysis routine would primarily involve copying the contents within this nesting from the development macro to the corresponding location within this nesting in ROI_analysis with as few additional changes elsewhere as possible. The primary difference in the analysis routine is the x/y_centers are no longer known prior to entering the nesting, they are parsed one at a time immediately before their values are needed, so all assignments/operations involving these arrays elsewhere must be removed. In addition, although the loop over each ROI_selection_radius is retained and this is only applicable to circle selections, its value is not used by any other ROI selection shape but the output data folder will still be named based on this value in the short term. It was decided that an implementation that avoids modification of the nesting/looping hierarchy would be the easiest to achive functionality with and subsequent commits can address this further.

As a side note, the various analysis/ROI definition config files are now stored in a subdirectory "AnalysisConfigs" separately from the various macros. This commit does not remove all moved files from their original locations, but a commit in the near future will focus on any remaining organizational conflicts and ensure the pCT_Tools repo is properly cleaned/organized.