Finished implementation of expanded ROI selection analysis scheme in ROI_analysis and configured ROI_analysis_control to operate it
The new ROI analysis scheme was essentially copied and pasted into ROI_analysis in place of the previous scheme and all plot related tasks were removed ( a backup with the plots still included was created as ROI_analysis_plotting) to simplify the process. Once it was confirmed that this ROI_anlysis implementation was working properly, ROI_analysis_control was updated to be able to pass analysis target directories to ROI_analysis. Despite some difficulties, particularly with all the new variables/functions developed for the new ROI analysis scheme missing from ROI_analysis_control, it now appears to be working properly and the first round of analyses are currently being performed.