Fixed small bug in Results table indexing and identified/resolved an issue with shaded ROI selections w/ duplicate ROI selections; also added a few bash tools applicable to pCT data manipulation
The bash tools added in this commit are for copying of TV and ROI analysis data files from their reconstruction_data directory to the current directory with their names changed to reflect some aspect of their reconstruction parameters so they can exist in a single directory together while preserving the information identifying the reconstruction they come from instead of the name of the directory in which they were originally located providing this info. Tools were also added to remove the _n suffix added to a reconstruction data directory when the (n-1) directory already exists (where _0 appears without either of these characters by convention). This is necessary because sometimes an error occurs in either reconstruction or the transferring of data to Kodiak or a home computer that causes a number of empty reconstruction data directories so subsequent reconstructions will add _n to the name instead of writing to the empty directory. This happened for every TV=1 reconstruction for the HN715 phantom so tools were created to make it possible to remove the _1 from all of these directory names. These were created for this particular instance and another tool was previously created to change _n to _m for any n/m, but these tools applied to every subdirectory inside a directory and the new tools allows the user to provide a substring to match as well. The new function is only intended to remove _1 from all directories matching the specified string, but this will later be extended to allow for any _n and will likely be combined with the previously created functions to make a single function that performs any desired manipulation/renaming/deleting of these directories whose names are assigned according to the create_unique_dir C++ function inside the reconstruction program.

ROI analysis of an entire set of reconstructions using multiple varying parameter values and their resulting parameter value combinations have now been analyzed using ROI_analysis_control to determine and then iteratively pass the target reconstruction directory to ROI_analysis so that the entire process proceeds automatically without user intervention with a single execution of the control macro. A brief inspection of the resulting output analysis data confirms that the new ROI selection and analysis scheme is working properly and is able to be called externally from another macro as desired.

The writing of the Results table to CSV was still configured to assume each ROI appeared in every slice to be analyzed, so the CSV ended up with rows of zeros for each ROI that wasn't analyzed in a particular slice. The analysis function now receives an additional argument specifying the number of ROIs that were analyzed in the current slice and this value is provided by an ROI counter variable reset to zero at the beginning of each slice loop and incremented each time an ROI is analyzed within this loop, resulting in this counter holding the total count of ROIs analyzed in a particular slice at the end of the loop where the CSV file is written. Hence, CSV files are now written correctly.

This commit also addresses an issue with the ROI selection shading/labeling scheme. Technically the filling of an ROI selection with white color changes the values of the voxels within the selection to the value of the color white. Since the fill occurs after the measurements within the ROI are taken, this usually does not cause an issue. However, since the Tooth region of the HN715 phantom contains both dentin and enamel, this ROI is analyzed once using each of these materials as the theoretical material for error calculations, so the 2nd time this ROI is selected, the mean RSP measurement is no longer the same since the white filling changes the voxel values. Hence, this shading routine was turned off for the current analysis. A number of shaded/labeled ROI selection image slice PNGs were already created so one of these can be used as the image to convery this information, but this is not needed for every single reconstruction and with this now turned off the mean measured RSP is the same for both ROI selections as expected.

The ROI definition config file for the CTP404 phantom has now been modified for the new ROI selection scheme with the circular selections of the original implementation replaced by the "oval" ROI selection even though the new scheme includes support for the original "circular_ROI" selection routine (just renamed to "circle" now). Once the original analysis can be equivalently performed using the new scheme, support for the original scheme will be removed since the process is easier, faster, and more reliable with the new scheme and the code can then be simplified as the original implementation requires considerably more work and code than the new scheme. This will also allow several field lines to be removed from the configuration file since the information provided in these lines will then be specified in the x/yparams lines in a clearer and more consise way.

Now that the automated analysis of a large set of reconstructions has been performed with the new ROI analysis scheme, the next step is to merge these results into single files with the merge_data and merge_data_control macros. Similar to the issue encountered with the lines of zeros in the CSV files due to each ROI not being analyzed in every slice in which 1 or more ROIs are analyzed, the parsing of these CSVs and merging of their data together will likely require some modifications to the merge_data macro. Unfortunately, such changes may be fairly substantive, but it is unclear at this point how problematic this will be. Most likely this will have to include the addition of the image/execution and ROI definition configuration file parsing and this should make the remaining modifications similar to those made to ROI_analysis. However, these config file parsing schemes rely on and require a large # of variable and function definitions and a number of conflicts with these were encountered when these routines were added to ROI_analysis, so hopefully this process is not as problematic as that one was and primarily requires the ADDITION of variables/functions to the macro from one of the others and not the modification of functions or variable values/names as this is a consideraby more time consuming process. The most unfortunate aspect of this development is the fact that each macro has a very small allocation of memory, so all variables/functions from the analysis macros cannot simply be added to the merge_data macro in addition to those already there as this would undoubtedly exceed the memory allocation, as was previously encountered when all variables/functions in ROI_analysis and development_testbench were attempted to be combined and execution would not even reach the end of variable definitions and beginning of the actual analysis portion of the macros. However, once this merge_data macro is working with the new ROI analysis output data, this entire set of pCT reconstruction analysis macros will finally be finished and will now run quite quickly (4-5 reconstructions can be analyzed per minute and merging of the analysis data previously took minutes to complete with 400+ reconstructions as input, which won't change significantly regardless of how much work it takes to get functioning as the computational load and complexity will not change.