Update of ImageJ w/ support for new expanded ROI selection analysis and updated pCT user profile/tools
The merge_data macro has been modified to support the output data from usage of the new ROI selection analysis scheme. Initial testing appears to confirm that the modified merge_data implementation works.

This also includes an update of the WinSCP configuration file for pCT users and some modifications and additions to the bash profile and associated tools.